Renaming AGauge control to AGaugeControl
Help with namespace/class confusion. Planning on modularizing the AGauge control further, so more breaking changes ahead.
- Removed static copyright header in some code files in favor of root-level copyright
