+++ AGauge/AGaugeLabel.cs
@@ -1,29 +1,4 @@
-﻿// Copyright (C) 2007 A.J.Bauer
-//
-//  This software is provided as-is, without any express or implied
-//  warranty.  In no event will the authors be held liable for any damages
-//  arising from the use of this software.
-
-//  Permission is granted to anyone to use this software for any purpose,
-//  including commercial applications, and to alter it and redistribute it
-//  freely, subject to the following restrictions:
-//  1. The origin of this software must not be misrepresented; you must not
-//     claim that you wrote the original software. if you use this software
-//     in a product, an acknowledgment in the product documentation would be
-//     appreciated but is not required.
-//  2. Altered source versions must be plainly marked as such, and must not be
-//     misrepresented as being the original software.
-//  3. This notice may not be removed or altered from any source distribution.
-//
-// -----------------------------------------------------------------------------------
-// Copyright (C) 2012 Code Artist
-// 
-// Added several improvement to original code created by A.J.Bauer.
-// Visit: http://codearteng.blogspot.com for more information on change history.
-//
-// -----------------------------------------------------------------------------------
-
-using System;
+﻿using System;
 using System.Drawing;
 
 namespace AGauge
@@ -38,8 +13,8 @@ namespace AGauge
     private string _Name;
 
     [System.ComponentModel.Browsable(false)]
-    public void SetOwner(Controls.AGauge value) { Owner = value; }
-    private Controls.AGauge Owner;
+    public void SetOwner(AGaugeControl value) { Owner = value; }
+    private AGaugeControl Owner;
 
     private void NotifyOwner() { if (Owner != null) Owner.RepaintControl(); }
 
@@ -47,7 +22,7 @@ namespace AGauge
     {
       if (Owner != null)
       {
-        Owner.NotifyChanging(nameof(Controls.AGauge.GaugeLabels));
+        Owner.NotifyChanging(nameof(AGaugeControl.GaugeLabels));
       }
     }
 
@@ -55,7 +30,7 @@ namespace AGauge
     {
       if (Owner != null)
       {
-        Owner.NotifyChanged(nameof(Controls.AGauge.GaugeLabels));
+        Owner.NotifyChanged(nameof(AGaugeControl.GaugeLabels));
       }
     }
